% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{ajustar_bn}
\alias{ajustar_bn}
\title{Ajuste de una Red Bayesiana a un set de datos}
\usage{
ajustar_bn(network, data)
}
\arguments{
\item{network}{network definida con \code{\link{definir_bn}}.}

\item{data}{tibble con los datos (ver \code{\link{cargar_datos}}).}
}
\description{
Ajuste de una Red Bayesiana a un set de datos
}
