% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{ajustar_bn_boot}
\alias{ajustar_bn_boot}
\title{Ajuste de Red Bayesianas a una lista de bootstraped data}
\usage{
ajustar_bn_boot(network, bootstraped_data)
}
\arguments{
\item{network}{network definida con \code{\link{definir_bn}}.}

\item{data}{una lista de tibble de acuerdo a \code{\link{bootstrap_data}}.}
}
\description{
Ajuste de Red Bayesianas a una lista de bootstraped data
}
