% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Genera un resampling basado en bootstrap}
\usage{
bootstrap_data(my_data, nodes, n_samples = 3, n_boots = 100)
}
\arguments{
\item{my_data}{tibble con los datos (ver \code{\link{cargar_datos}}).}

\item{nodes}{puntos de muestreo a utilizar}

\item{n_samples}{numero de repeticiones a incluir en cada bootstrap sample.}

\item{n_boots}{numero de bootstrap samples.}
}
\value{
Una lista en que cada miembro es un tibble con los resultados de cada
bootstrap sample.
}
\description{
Genera un resampling basado en bootstrap
}
