% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_visualizacion.R
\name{cajas_y_bigotes}
\alias{cajas_y_bigotes}
\title{Grafico de cajas y bigotes}
\usage{
cajas_y_bigotes(
  my_data,
  resaltar_sesgados = FALSE,
  orden_x = NULL,
  eje_x = NULL,
  eje_y = NULL
)
}
\arguments{
\item{my_data}{Tibble con los datos (ver \code{\link{cargar_datos}}).}

\item{resaltar_sesgados}{Mostrar los valores sesgados como rombos rojos (FALSE by default).}

\item{orden_x}{Vector describiendo el orden de las cajas (alfabeticamente by default).}

\item{eje_x}{Nombre del eje x.}

\item{eje_y}{Nombre del eje y.}
}
\description{
Grafico de cajas y bigotes
}
