% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_datos.R
\name{cargar_datos}
\alias{cargar_datos}
\title{Cargar datos de un Excel para el análisis}
\usage{
cargar_datos(archivo, hoja)
}
\arguments{
\item{archivo}{Lugar donde esta guardado el fichero Excel. Se recomienda usar
paths relativos.}

\item{hoja}{Nombre de la hoja Excel donde estan los datos.}
}
\value{
Un tibble con 5 columnas: fecha, planta, punto, recuento, rec_minimo.
}
\description{
Cargar datos de un Excel para el análisis
}
\details{
El archivo de Excel debe tener 5 columnas en el siguiente orden:
\itemize{
\item{Fecha de la muestra}
\item{Identificador de la planta (EDAR, localización...)}
\item{Identificador del punto de muestreo (Primario, PE...)}
\item{Recuento microbioano. Se recomienda usar CFU/100 mL. Si no, habrá que hacer
los cambios de unidades relevantes en las funciones de simulación.}
\item{En el caso de recuentos sesgados (al menos X), el valor.}
}
}
