% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_datos.R
\name{codigo_ausencia}
\alias{codigo_ausencia}
\title{Asigna el valor 0 a los datos de ausencia codificados con otro valor}
\usage{
codigo_ausencia(my_data, code_abscence)
}
\arguments{
\item{my_data}{Tibble con los datos (ver \code{\link{cargar_datos}}).}

\item{code_abscence}{Codigo utilizado en el Excel para los datos de ausencia}
}
\description{
Asigna el valor 0 a los datos de ausencia codificados con otro valor
}
