% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{definir_bn}
\alias{definir_bn}
\title{Define una network en base a una vector de nodos}
\usage{
definir_bn(nodes)
}
\arguments{
\item{nodes}{un vector con los nombres de los nodos

Define un grafo recto en que cada nodo es hijo del anterior.}
}
\description{
Define una network en base a una vector de nodos
}
