% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_Roland.R
\name{Ec_Leafy2}
\alias{Ec_Leafy2}
\title{Modelo de E.coli en planta}
\usage{
Ec_Leafy2(Days, P_Ec_Water, C_Ecoli_Irr_Water, Rain_Days, sun_hrs_mode)
}
\arguments{
\item{Days}{numero de dias incluido en la simulacion.}

\item{P_Ec_Water}{prevalencia de Ecoli en el agua (usar 1 si C_Ecoli_Irr_Water ya incluye prevalencia).}

\item{C_Ecoli_Irr_Water}{Concentración de E.coli en el agua de irrigación (en CFU per ml of water).}

\item{sun_hrs_mode}{Moda del número de horas de sol por día.}

\item{Rain_days}{Numero de días de lluvia durante la simulación.}
}
\description{
Modelo de E.coli en planta
}
