% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{extraer_pars_de_bn}
\alias{extraer_pars_de_bn}
\title{Extrae los parámetros de una lista de Redes Bayesianas}
\usage{
extraer_pars_de_bn(my_nets)
}
\arguments{
\item{my_nets}{una lista de redes bayesianas de acuerdo a \code{\link{ajustar_bn_boot}}.}
}
\description{
Extrae los parámetros de una lista de Redes Bayesianas
}
