% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_visualizacion.R
\name{histograma_punto}
\alias{histograma_punto}
\title{Muestra el histograma del recuento en un punto de muestreo}
\usage{
histograma_punto(my_data, sample, eje_x = NULL, eje_y = NULL)
}
\arguments{
\item{my_data}{tibble con los datos (ver \code{\link{cargar_datos}}).}

\item{sample}{punto de muestra.}

\item{eje_x}{nombre del eje x}

\item{eje_y}{nombre del eje  y}
}
\description{
Muestra el histograma del recuento en un punto de muestreo
}
