% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{parametros_ZIP}
\alias{parametros_ZIP}
\title{Extrae los parametros de un modelo ZIP}
\usage{
parametros_ZIP(ZIP_model)
}
\arguments{
\item{ZIP_model}{un modelo ZIP (ver \code{\link{ajustar_ZIP}}).}
}
\description{
Extrae los parametros de un modelo ZIP
}
