% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{plot_parametros_bn}
\alias{plot_parametros_bn}
\title{Muestra la incertidumbre de los parametros usando cajas y bigotes}
\usage{
plot_parametros_bn(parametros_bn, orden)
}
\arguments{
\item{parametros_bn}{el resultado de \code{\link{extraer_pars_de_bn}}.}

\item{orden}{orden de los parametros}
}
\description{
Muestra la incertidumbre de los parametros usando cajas y bigotes
}
