% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_BN.R
\name{resumen_parametros_bn}
\alias{resumen_parametros_bn}
\title{Resumen de la variacion de los parametros de una bootstrapped BN}
\usage{
resumen_parametros_bn(parametros_bn)
}
\arguments{
\item{parametros_bn}{el resultado de \code{\link{extraer_pars_de_bn}}.}
}
\description{
Resumen de la variacion de los parametros de una bootstrapped BN
}
