% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_datos.R
\name{sesgados_por_mediana}
\alias{sesgados_por_mediana}
\title{Sustituye los datos sesgados por la mediana del punto de muestreo}
\usage{
sesgados_por_mediana(my_data)
}
\arguments{
\item{my_data}{Tibble con los datos (ver \code{\link{cargar_datos}}).}
}
\description{
Sustituye los datos sesgados por la mediana del punto de muestreo
}
