% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_exposicion.R
\name{simular_riego}
\alias{simular_riego}
\title{Simulacion de la exposicion en planta}
\usage{
simular_riego(
  ZIP_model,
  Days,
  Rain_Days,
  sun_hrs_mode,
  irrigation_type = "normal",
  nsims = 1000
)
}
\arguments{
\item{ZIP_model}{un modelo ZIP (ver \code{\link{ajustar_ZIP}}).}

\item{Days}{numero de dias incluido en la simulacion.}

\item{sun_hrs_mode}{Moda del número de horas de sol por día.}

\item{irrigation_type}{Tipo de irrigacion (normal, goteo o noSplashing).}

\item{Rain_days}{Numero de días de lluvia durante la simulación.}

\item{n_sims}{numero de simulaciones de Monte Carlo (1000 by default).}
}
\value{
Un tibble con tres columnas: day, count (recuento microbiano en CFU/mL)
and sim (numero de la simulacion de Monte Carlo).
}
\description{
Simulacion de la exposicion en planta
}
