% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIP.R
\name{simular_ZIP}
\alias{simular_ZIP}
\title{Simulación de una ZIP usando Monte Carlo}
\usage{
simular_ZIP(model, n_sims = 1000)
}
\arguments{
\item{n_sims}{numero de simulaciones de Monte Carlo (1000 by default).}

\item{ZIP_model}{un modelo ZIP (ver \code{\link{ajustar_ZIP}}).}
}
\description{
Simulación de una ZIP usando Monte Carlo
}
